feat: simplest demo for debugging

+++ frb_example/flutter_via_integrate/rust/src/lib.rs
@@ -1,3 +1,3 @@
-pub fn add(left: usize, right: usize) -> usize {
+pub fn add(left: i32, right: i32) -> i32 {
     left + right
 }
